add to cart with cookie

+++ CoreUI/Controllers/CartController.cs
@@ -7,6 +7,8 @@ using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Identity;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.EntityFrameworkCore;
+using Newtonsoft.Json;
+using static CoreUI.Controllers.AddToCartWithCookie;
 
 namespace CoreUI.Controllers
 {
@@ -20,38 +22,60 @@ namespace CoreUI.Controllers
             _cartService = cartService;
             _userManager = userManager;
         }
-         public IActionResult Index()
+        public IActionResult Index()
         {
+
             var cart = _cartService.GetCartByUserId(_userManager.GetUserId(User));
-            return View(new CartModel(){
+            return View(new CartModel()
+            {
                 CartId = cart.Id,
-                CartItems = cart.CartItems.Select(i=>new CartItemModel()
+                CartItems = cart.CartItems.Select(i => new CartItemModel()
                 {
                     CartItemId = i.Id,
                     ProductId = i.ProductNavigation.Id,
                     Name = i.ProductNavigation.Brand,
                     Price = (double)i.ProductNavigation.Price,
                     ImageUrl = i.ItemPic.LaptopPicPath,
-                    Quantity =(int)i.Quantity
+                    Quantity = (int)i.Quantity
 
                 }).ToList()
             });
-        } 
+        }
 
         [HttpPost]
-        public IActionResult AddToCart(int productId,byte quantity,int itemPicId)
+        public IActionResult AddToCart(int productId, byte quantity, int itemPicId)
         {
-            var userId = _userManager.GetUserId(User);
-            _cartService.AddToCart(userId,productId,quantity,itemPicId);
-            return RedirectToAction("Index");
-        } 
+            if (User.Identity.IsAuthenticated)
+            {
+                var userId = _userManager.GetUserId(User);
+                _cartService.AddToCart(userId, productId, quantity, itemPicId);
+                return RedirectToAction("Index");
+            }
+            else
+            {
+                CoreDbContext db = new CoreDbContext();
+                var data = db.Laptops.FirstOrDefault(i => i.Id == productId);
+                HttpContext.Response.Cookies.Append("key", JsonConvert.SerializeObject(data));
+
+
+
+                return RedirectToAction("TestView", "AddToCartWithCookie");
+            }
+        }
 
         [HttpPost]
         public IActionResult DeleteFromCart(int productId)
         {
-            var userId = _userManager.GetUserId(User);
-            _cartService.DeleteFromCart(userId, productId);
-            return RedirectToAction("Index");
+            if (User.Identity.IsAuthenticated)
+            {
+                var userId = _userManager.GetUserId(User);
+                _cartService.DeleteFromCart(userId, productId);
+                return RedirectToAction("Index");
+            }
+            else
+            {
+               return RedirectToAction("Index");
+            }
         }
     }
 }
